feat : .env파일 추가 refactor : application.properties, main 이름 변경, 패키지 명 변경
