+++ .env
@@ -1,0 +1,3 @@
+DB_URL=jdbc:mariadb://localhost:3306/letmovie
+DB_USERNAME=root
+DB_PASSWORD=!123456
