+++ src/main/java/com/example/letmovie/domain/board/entity/Comment.java
@@ -1,0 +1,4 @@
+package com.example.letmovie.domain.board.entity;
+
+public class Comment {
+}
