+++ src/main/java/com/example/letmovie/domain/movie/entity/Showtime.java
@@ -1,0 +1,29 @@
+package com.example.letmovie.domain.movie.entity;
+
+import com.example.letmovie.domain.reservation.entity.Screen;
+import jakarta.persistence.*;
+import lombok.Getter;
+
+import java.util.Date;
+
+@Entity
+@Getter
+public class Showtime {
+
+    @Id
+    @Column(name = "showtime_id")
+    private Long id;
+
+    @ManyToOne
+    @JoinColumn(name="screen_id")
+    private Screen screen;
+
+    @ManyToOne
+    @JoinColumn(name = "movie_id")
+    private Movie movie;
+
+    @Column(nullable = false)
+    private Date showtimeDate;
+    @Column(nullable = false)
+    private Date showtimeTime;
+}
