+++ src/main/java/com/example/letmovie/domain/movie/entity/Status.java
@@ -1,0 +1,6 @@
+package com.example.letmovie.domain.movie.entity;
+
+public enum Status {
+
+    PREV, SHOW, END
+}
