+++ src/main/java/com/example/letmovie/domain/payment/entity/Payment.java
@@ -1,0 +1,4 @@
+package com.example.letmovie.domain.payment.entity;
+
+public class Payment {
+}
