+++ src/main/java/com/example/letmovie/domain/reservation/entity/Reservation.java
@@ -1,0 +1,6 @@
+package com.example.letmovie.domain.reservation.entity;
+
+public class Reservation {
+
+
+}
