+++ src/main/java/com/example/letmovie/domain/reservation/entity/ReservationSeat.java
@@ -1,0 +1,4 @@
+package com.example.letmovie.domain.reservation.entity;
+
+public class ReservationSeat {
+}
