--- src/main/java/com/example/nbe232team2/Nbe232Team2Application.java
@@ -1,13 +1,0 @@
-package com.example.nbe232team2;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Nbe232Team2Application {
-
-    public static void main(String[] args) {
-        SpringApplication.run(Nbe232Team2Application.class, args);
-    }
-
-}
