--- src/main/java/com/example/nbe232team2/domain/member/entity/Grade.java
@@ -1,6 +1,0 @@
-package com.example.nbe232team2.domain.member.entity;
-
-public enum Grade {
-    GENERAL, VIP
-
-}
