--- src/main/java/com/example/nbe232team2/domain/movie/entity/Status.java
@@ -1,6 +1,0 @@
-package com.example.nbe232team2.domain.movie.entity;
-
-public enum Status {
-
-    PREV, SHOW, END
-}
