--- src/main/java/com/example/nbe232team2/domain/payment/entity/Payment.java
@@ -1,4 +1,0 @@
-package com.example.nbe232team2.domain.payment.entity;
-
-public class Payment {
-}
