--- src/main/java/com/example/nbe232team2/domain/reservation/entity/Screen.java
@@ -1,14 +1,0 @@
-package com.example.nbe232team2.domain.reservation.entity;
-
-import jakarta.persistence.*;
-
-@Entity
-public class Screen {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "screen_id")
-    private Long id;
-
-
-}
