--- src/main/java/com/example/nbe232team2/global/config/SwaggerConfig.java
@@ -1,24 +1,0 @@
-package com.example.nbe232team2.global.config;
-
-import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.info.Info;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class SwaggerConfig {
-
-    @Bean
-    public OpenAPI openAPI() {
-        return new OpenAPI()
-                .info(apiInfo());
-    }
-
-    private Info apiInfo() {
-
-        return new Info()
-                .title("LetMovie")
-                .description("2차 프로젝트")
-                .title("v1");
-    }
-}
