+++ src/main/resources/application.properties
@@ -1,9 +1,13 @@
 spring.application.name=NBE2-3-2-team2
 
+# Import .env file
+spring.config.import=optional:file:.env[.properties]
+
+# Mariadb
+spring.datasource.driver-class-name=org.mariadb.jdbc.Driver
+
+spring.datasource.url=${DB_URL}
+spring.datasource.username=${DB_USERNAME}
+spring.datasource.password=${DB_PASSWORD}
 
 spring.jpa.hibernate.ddl-auto=create
-spring.datasource.driverClassName=org.mariadb.jdbc.Driver
-spring.datasource.url="jdbc:mariadb://localhost:3306/letmovie"
-spring.datasource.username="root"
-spring.datasource.password="!123456"
-
