+++ src/test/java/com/example/letmovie/LetMovieApplicationTests.java
@@ -1,0 +1,13 @@
+package com.example.letmovie;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class LetMovieApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
